% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\name{coastline}
\alias{coastline}
\title{Get coastline data}
\usage{
coastline(extent = NULL, ..., dimension = c(1024, 1024), projection = NULL)
}
\arguments{
\item{extent}{extent xmin, xmax, ymin, ymax}

\item{...}{argumetns passed to [elevation()]}

\item{dimension}{optional size of matrix to contour (a default is used)}

\item{projection}{provide this for custom extents in a given projection}
}
\value{
matrix of coordinates in longlat (or whatever projection is specified)
}
\description{
Get a coatline for any region on Earth.
}
\details{
This uses the [elevation()] function to get a coastline by contouring elevation data at 0. (This might return nothing.)

For 'extent' not provided it will be derived from a plot extent if one exists. It's the user's job to ensure the
projection is appropriate (we don't have a graphics mechanism to record a current projection string.)

If no device is open and no extent given, a coastline for longlat is given.
}
\examples{
plot(coastline(extent = c(-180, 180, -90, 90)), type = "l")

}
