% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatarelief.R
\name{elevation}
\alias{elevation}
\title{Get digital elevation data}
\usage{
elevation(
  extent = c(-180, 180, -90, 90),
  ...,
  dimension = NULL,
  projection = NULL,
  resample = "bilinear",
  source = NULL,
  threshold = 0.5
)
}
\arguments{
\item{extent}{a numeric vector of xmin,xmax,ymin,ymax or a terra or raster rast object}

\item{...}{arguments passed to 'vapour::vapour_warp_raster'}

\item{dimension}{optional output size in ncol,nrow (x,y)}

\item{projection}{optional coordinate reference system to use (map projection), longlat is assumed}

\item{resample}{resampling algorithm for the GDAL warper, "bilinear" by default}

\item{source}{a GDAL raster source, to override the inbuild GEBCO + SRTM (in future we might patch in local source)}

\item{threshold}{a size in degrees above which no SRTM data is queried (about )}
}
\value{
a matrix, or depending on format of 'extent' a terra rast or raster object with elevation data
}
\description{
Read elevation data for any region on Earth.
}
\details{
Originally using GEBCO 2019 as a background, and SRTM 30m resolution for
regions that fit approximately within the size of an SRTM tile (these are 1 degree wide).

Note that data is streamed into memory, so don't make the dimensions of the 'x' target raster too big.

To use these data, please attribute the use of GEBCO to the  GEBCO
Compilation Group (2019) GEBCO 2019 Grid
(doi:10.5285/836f016a-33be-6ddc-e053-6c86abc0788e), and the use of NASADEM to
NASA JPL. NASADEM Merged DEM Global 1 arc second V001. 2020, distributed by
NASA EOSDIS Land Processes DAAC,
https://doi.org/10.5067/MEaSUREs/NASADEM/NASADEM_HGT.001.
}
\examples{

image(elevation())
image(elevation(c(100, 150, -60, -20)))
elevation(terra::rast())
elevation(raster::raster())

elevation(raster::raster(raster::extent(80, 120, -60, -40), res = 0.25, crs = "OGC:CRS84"))

elevation(raster::raster(raster::extent(c(-1, 1, -1, 1) * 15e3), nrows = 1024, ncols = 1024,
   crs = "+proj=laea +lat_0=44.6371 +lon_0=-63.5923"))
}
