% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatarelief.R
\name{find_relief}
\alias{find_relief}
\title{Find files (not functional, just a stub)}
\usage{
find_relief(pattern = "etopo2.*nc$")
}
\arguments{
\item{pattern}{}
}
\description{
Find files (not functional, just a stub)
}
