% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{gebco}
\alias{gebco}
\alias{gebco22}
\alias{gebco21}
\alias{gebco19}
\title{GEBCO source dsn}
\usage{
gebco(vsi = TRUE)

gebco21(vsi = TRUE)

gebco22(vsi = TRUE)

gebco19(vsi = TRUE)
}
\arguments{
\item{vsi}{include the 'vsicurl' prefix (`TRUE` is default)}
}
\value{
character string, URL to online GeoTIFF
}
\description{
A data source name to the GEBCO  elevation 'COG' GeoTIFF.
}
\details{
GEBCO 2022 is created and hosted by Philippe Massicotte. 

GEBCO 2019 and 2021 created and hosted by the Australian Antarctic Division.
}
\examples{
gebco()
}
