% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{gebco21}
\alias{gebco21}
\title{GEBCO 2021 source dsn}
\usage{
gebco21(vsi = TRUE)
}
\arguments{
\item{vsi}{include the 'vsicurl' prefix (`TRUE` is default)}
}
\value{
character string, URL to online GeoTIFF
}
\description{
A data source name to the GEBCO 2021 elevation 'COG' GeoTIFF.
}
\examples{
gebco21()
}
