% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatarelief.R
\name{get_relief}
\alias{get_relief}
\title{Get a bowerbird source}
\usage{
get_relief(src, local_file_root = NULL)
}
\arguments{
\item{src}{bowerbird source (assumed to be a topo source)}

\item{local_file_root}{where to put the bowerbird loot}
}
\description{
This is very raw, see https://ropensci.org/blog/2018/11/13/antarctic/
This could house a set of sources of known topography, and hide the bowerbird/blueant stuff.
}
