% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_seaice.R
\name{raad_source}
\alias{raad_source}
\title{Data sources raadtools knows about}
\usage{
raad_source(id = "nsidc_25km_seaice", ...)
}
\arguments{
\item{id}{raadtools id of data sourced}

\item{...}{unused}
}
\value{
data frame of dates and vrt dsn (names vary, this is WIP)
}
\description{
Currently just 'nsidc_25km_seaice' is available.
}
\details{
Obtains a listing of GDAL-read URLs indexed by raadtools.

I think the early URLs are wrong ... but later ones work, see the README
}
\examples{
raad_source("nsidc_25km_seaice")
}
