% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatarelief.R
\name{relief_patcher}
\alias{relief_patcher}
\title{Crop raster data to each polygon}
\usage{
relief_patcher(polygons, relief, pname = NULL)
}
\arguments{
\item{polygons}{Spatial Polygons}

\item{relief}{topo raster}

\item{pname}{project name}
}
\value{
list of rasters (GeoTIFFs on disk)
}
\description{
Crop raster data to each polygon
}
