% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatarelief.R
\name{whatarelief_bb_cache}
\alias{whatarelief_bb_cache}
\title{bowerbird cache for whatarelief}
\usage{
whatarelief_bb_cache()
}
\value{
file path to bowerbird goodies
}
\description{
bowerbird cache for whatarelief
}
\examples{
whatarelief_bb_cache()
}
