% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatarelief.R
\name{whatarelief_cache}
\alias{whatarelief_cache}
\title{whatarelief cache for project tiles}
\usage{
whatarelief_cache()
}
\value{
file path to project home
}
\description{
whatarelief cache for project tiles
}
\examples{
whatarelief_cache()
}
