% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagery.R
\name{imagery}
\alias{imagery}
\title{Get imagery data}
\usage{
imagery(
  extent = c(-180, 180, -90, 90),
  ...,
  dimension = NULL,
  projection = "OGC:CRS84",
  resample = "bilinear",
  source = NULL
)
}
\arguments{
\item{extent}{a numeric vector of xmin,xmax,ymin,ymax or a terra or raster rast object}

\item{...}{arguments passed to 'vapour::vapour_warp_raster'}

\item{dimension}{optional output size in ncol,nrow (x,y)}

\item{projection}{optional coordinate reference system to use (map projection), longlat is assumed}

\item{resample}{resampling algorithm for the GDAL warper, "near" by default}

\item{source}{a GDAL raster source, to override the inbuilt source/s}
}
\value{
a matrix, or depending on format of 'extent' a terra rast or raster object with elevation data
}
\description{
Read imagery data for (almost?) any region on Earth.
}
\details{
Originally using OSM.

Note that data is streamed into memory, so don't make the dimensions of the 'x' target raster too big.

To use these data, please attribute
}
\examples{

plt <- function(x) {
nr <- nrow(x)
nc <- ncol(x)
plot(NA, xlim = c(0, nc), ylim = c(0, nr), asp = 1);
 rasterImage(x, 0, 0, nc, nr)
 }

plt(imagery())
plt(imagery(c(100, 150, -60, -20), projection = "OGC:CRS84"))
## can't do yet
#plt(imagery(terra::rast())
#imagery(raster::raster())

}
